fix: lint issue resolve
