+++ monorepo/apps/example/src/pages/Elixir/index.tsx
@@ -1,5 +1,5 @@
-import React from 'react';
 import { Elixir } from '@pangolindex/elixir';
+import React from 'react';
 
 function ElixirUI() {
   return (
