+++ monorepo/packages/elixir/src/components/Elixir/styles.tsx
