+++ monorepo/packages/elixir/src/hooks/chart/evm.ts
@@ -11,11 +11,11 @@ import {
 import { TickData, Ticks, ZERO_ADDRESS, useAllV3TicksQuery, useChainId, wrappedCurrency } from '@pangolindex/shared';
 import { useSingleContractMultipleData } from '@pangolindex/state-hooks';
 import { useCallback, useEffect, useMemo, useState } from 'react';
-import { useTickLensContract } from 'src/utils/contracts';
-import computeSurroundingTicks from 'src/utils/computeSurroundingTicks';
 import { ChartEntry } from 'src/components/LiquidityChartRangeInput/types';
 import { usePool } from 'src/hooks/common';
 import { PoolState } from 'src/hooks/types';
+import computeSurroundingTicks from 'src/utils/computeSurroundingTicks';
+import { useTickLensContract } from 'src/utils/contracts';
 import { TickProcessed } from './types';
 
 export function useDensityChartData({
