+++ monorepo/packages/shared/src/utils/common.ts
