+++ monorepo/packages/state-hooks/src/hooks/index.tsx
@@ -13,8 +13,8 @@ import useParsedQueryString from 'src/hooks/useParsedQueryString';
 import useTransactionDeadline from 'src/hooks/useTransactionDeadline';
 import { useUSDCPriceHook } from 'src/hooks/useUSDCPrice';
 import { useUSDCPrice } from 'src/hooks/useUSDCPrice/evm';
-import { usePair } from './usePair/evm';
 import { useToken } from './tokens/evm';
+import { usePair } from './usePair/evm';
 import { useTotalSupplyHook } from './useTotalSupply';
 
 // hooks
