+++ monorepo/packages/swap/package.json
@@ -56,6 +56,7 @@
     "@ethersproject/providers": "5.6.8",
     "@gelatonetwork/limit-orders-react": "2.8.2",
     "@pangolindex/core": "*",
+    "@pangolindex/elixir": "*",
     "@pangolindex/token-drawer": "*",
     "@pangolindex/shared": "*",
     "@pangolindex/state-hooks": "*",
