+++ monorepo/packages/swap/src/hooks/Trades.ts
@@ -1,24 +1,19 @@
+import { PoolState, usePoolsHook } from '@pangolindex/elixir';
 import {
   Currency,
   CurrencyAmount,
-  // TODO: when add elixir package
-  // ElixirPool,
-  // ElixirTrade,
-  // FeeAmount,
+  ElixirPool,
+  ElixirTrade,
+  FeeAmount,
   Pair,
   Percent,
   Token,
   Trade,
 } from '@pangolindex/sdk';
-import { PairState, useChainId, wrappedCurrency, BIPS_BASE } from '@pangolindex/shared';
+import { BIPS_BASE, PairState, useChainId, wrappedCurrency } from '@pangolindex/shared';
 import { usePairsHook } from '@pangolindex/state-hooks';
 import flatMap from 'lodash.flatmap';
-import { useMemo } from 'react';
-
-// TODO: when add elixir package
-// import { usePoolsHook } from 'src/hooks/elixir/hooks';
-// TODO: when add elixir package
-// import { PoolState } from 'src/hooks/elixir/hooks/types';
+import { useEffect, useMemo, useState } from 'react';
 import { BASES_TO_CHECK_TRADES_AGAINST, CUSTOM_BASES } from 'src/constants';
 import { useDaasFeeInfo, useDaasFeeTo } from '../state/hooks/common';
 
@@ -145,162 +140,160 @@ export function useTradeExactOut(
     return { trade: null, isLoading: true };
   }, [allowedPairs, isLoading, currencyIn, currencyAmountOut, feeTo, feeInfo]);
 }
-// TODO: when add elixir package
-// function useAllElixirCommonPools(
-//   currencyA?: Currency,
-//   currencyB?: Currency,
-// ): {
-//   pools: ElixirPool[];
-//   isLoading: boolean;
-// } {
-//   const chainId = useChainId();
 
-//   const usePools = usePoolsHook[chainId];
+function useAllElixirCommonPools(
+  currencyA?: Currency,
+  currencyB?: Currency,
+): {
+  pools: ElixirPool[];
+  isLoading: boolean;
+} {
+  const chainId = useChainId();
 
-//   const bases: Token[] = chainId ? BASES_TO_CHECK_TRADES_AGAINST[chainId] : []; // eslint-disable-line react-hooks/exhaustive-deps
+  const usePools = usePoolsHook[chainId];
 
-//   const [tokenA, tokenB] = chainId
-//     ? [wrappedCurrency(currencyA, chainId), wrappedCurrency(currencyB, chainId)]
-//     : [undefined, undefined];
+  const bases: Token[] = chainId ? BASES_TO_CHECK_TRADES_AGAINST[chainId] : []; // eslint-disable-line react-hooks/exhaustive-deps
 
-//   const basePairs: [Token, Token][] = useMemo(
-//     () =>
-//       bases
-//         .flatMap((base): [Token, Token][] => bases.map((otherBase) => [base, otherBase]))
-//         // though redundant with the first filter below, that expression runs more often, so this is probably worthwhile
-//         .filter(([t0, t1]) => !t0.equals(t1)),
-//     [bases],
-//   );
+  const [tokenA, tokenB] = chainId
+    ? [wrappedCurrency(currencyA, chainId), wrappedCurrency(currencyB, chainId)]
+    : [undefined, undefined];
 
-//   const allCurrencyCombinations = useMemo(
-//     () =>
-//       tokenA && tokenB
-//         ? [
-//             // the direct pair
-//             [tokenA, tokenB] as [Token, Token],
-//             // token A against all bases
-//             ...bases.map((base): [Token, Token] => [tokenA, base]),
-//             // token B against all bases
-//             ...bases.map((base): [Token, Token] => [tokenB, base]),
-//             // each base against all bases
-//             ...basePairs,
-//           ]
-//             // filter out invalid pairs comprised of the same asset (e.g. WETH<>WETH)
-//             .filter(([t0, t1]) => !t0.equals(t1))
-//             // filter out duplicate pairs
-//             .filter(([t0, t1], i, otherPairs) => {
-//               // find the first index in the array at which there are the same 2 tokens as the current
-//               const firstIndexInOtherPairs = otherPairs.findIndex(([t0Other, t1Other]) => {
-//                 return (t0.equals(t0Other) && t1.equals(t1Other)) || (t0.equals(t1Other) && t1.equals(t0Other));
-//               });
-//               // only accept the first occurrence of the same 2 tokens
-//               return firstIndexInOtherPairs === i;
-//             })
-//             // optionally filter out some pairs for tokens with custom bases defined
-//             .filter(([tokenA, tokenB]) => {
-//               if (!chainId) return true;
-//               const customBases = CUSTOM_BASES[chainId];
+  const basePairs: [Token, Token][] = useMemo(
+    () =>
+      bases
+        .flatMap((base): [Token, Token][] => bases.map((otherBase) => [base, otherBase]))
+        // though redundant with the first filter below, that expression runs more often, so this is probably worthwhile
+        .filter(([t0, t1]) => !t0.equals(t1)),
+    [bases],
+  );
 
-//               const customBasesA: Token[] | undefined = customBases?.[tokenA.address];
-//               const customBasesB: Token[] | undefined = customBases?.[tokenB.address];
+  const allCurrencyCombinations = useMemo(
+    () =>
+      tokenA && tokenB
+        ? [
+            // the direct pair
+            [tokenA, tokenB] as [Token, Token],
+            // token A against all bases
+            ...bases.map((base): [Token, Token] => [tokenA, base]),
+            // token B against all bases
+            ...bases.map((base): [Token, Token] => [tokenB, base]),
+            // each base against all bases
+            ...basePairs,
+          ]
+            // filter out invalid pairs comprised of the same asset (e.g. WETH<>WETH)
+            .filter(([t0, t1]) => !t0.equals(t1))
+            // filter out duplicate pairs
+            .filter(([t0, t1], i, otherPairs) => {
+              // find the first index in the array at which there are the same 2 tokens as the current
+              const firstIndexInOtherPairs = otherPairs.findIndex(([t0Other, t1Other]) => {
+                return (t0.equals(t0Other) && t1.equals(t1Other)) || (t0.equals(t1Other) && t1.equals(t0Other));
+              });
+              // only accept the first occurrence of the same 2 tokens
+              return firstIndexInOtherPairs === i;
+            })
+            // optionally filter out some pairs for tokens with custom bases defined
+            .filter(([tokenA, tokenB]) => {
+              if (!chainId) return true;
+              const customBases = CUSTOM_BASES[chainId];
 
-//               if (!customBasesA && !customBasesB) return true;
+              const customBasesA: Token[] | undefined = customBases?.[tokenA.address];
+              const customBasesB: Token[] | undefined = customBases?.[tokenB.address];
 
-//               if (customBasesA && !customBasesA.find((base) => tokenB.equals(base))) return false;
-//               if (customBasesB && !customBasesB.find((base) => tokenA.equals(base))) return false;
+              if (!customBasesA && !customBasesB) return true;
 
-//               return true;
-//             })
-//         : [],
-//     [tokenA, tokenB, bases, basePairs, chainId],
-//   );
+              if (customBasesA && !customBasesA.find((base) => tokenB.equals(base))) return false;
+              if (customBasesB && !customBasesB.find((base) => tokenA.equals(base))) return false;
 
-//   const allCurrencyCombinationsWithAllFees: [Token, Token, FeeAmount][] = useMemo(
-//     () =>
-//       allCurrencyCombinations.reduce<[Token, Token, FeeAmount][]>((list, [tokenA, tokenB]) => {
-//         return list.concat([
-//           [tokenA, tokenB, FeeAmount.LOW],
-//           [tokenA, tokenB, FeeAmount.MEDIUM],
-//           [tokenA, tokenB, FeeAmount.HIGH],
-//         ]);
-//       }, []),
-//     [allCurrencyCombinations, chainId],
-//   );
+              return true;
+            })
+        : [],
+    [tokenA, tokenB, bases, basePairs, chainId],
+  );
 
-//   const pools = usePools(allCurrencyCombinationsWithAllFees);
+  const allCurrencyCombinationsWithAllFees: [Token, Token, FeeAmount][] = useMemo(
+    () =>
+      allCurrencyCombinations.reduce<[Token, Token, FeeAmount][]>((list, [tokenA, tokenB]) => {
+        return list.concat([
+          [tokenA, tokenB, FeeAmount.LOW],
+          [tokenA, tokenB, FeeAmount.MEDIUM],
+          [tokenA, tokenB, FeeAmount.HIGH],
+        ]);
+      }, []),
+    [allCurrencyCombinations, chainId],
+  );
 
-//   return useMemo(() => {
-//     return {
-//       pools: pools
-//         .filter((tuple): tuple is [PoolState.EXISTS, ElixirPool] => {
-//           return tuple[0] === PoolState.EXISTS && tuple[1] !== null;
-//         })
-//         .map(([, pool]) => pool),
-//       isLoading: pools.some(([state]) => state === PoolState.LOADING),
-//     };
-//   }, [pools]);
-// }
+  const pools = usePools(allCurrencyCombinationsWithAllFees);
 
-// TODO: when add elixir package
+  return useMemo(() => {
+    return {
+      pools: pools
+        .filter((tuple): tuple is [PoolState.EXISTS, ElixirPool] => {
+          return tuple[0] === PoolState.EXISTS && tuple[1] !== null;
+        })
+        .map(([, pool]) => pool),
+      isLoading: pools.some(([state]) => state === PoolState.LOADING),
+    };
+  }, [pools]);
+}
+
 /**
  * Returns the best trade for the exact amount of tokens in to the given token out
  */
-// export function useElixirTradeExactIn(
-//   currencyAmountIn?: CurrencyAmount,
-//   currencyOut?: Currency,
-// ): { trade: ElixirTrade | null; isLoading: boolean } {
-//   const [tradeData, setTradeData] = useState<{ trade: ElixirTrade | null; isLoading: boolean }>({
-//     trade: null,
-//     isLoading: true,
-//   });
+export function useElixirTradeExactIn(
+  currencyAmountIn?: CurrencyAmount,
+  currencyOut?: Currency,
+): { trade: ElixirTrade | null; isLoading: boolean } {
+  const [tradeData, setTradeData] = useState<{ trade: ElixirTrade | null; isLoading: boolean }>({
+    trade: null,
+    isLoading: true,
+  });
 
-//   const { pools: allowedPools, isLoading } = useAllElixirCommonPools(currencyAmountIn?.currency, currencyOut);
+  const { pools: allowedPools, isLoading } = useAllElixirCommonPools(currencyAmountIn?.currency, currencyOut);
 
-//   useEffect(() => {
-//     const getBestTradeExactIn = async () => {
-//       if (currencyAmountIn && currencyOut && allowedPools.length > 0 && !isLoading) {
-//         const trades = await ElixirTrade.bestTradeExactIn(allowedPools, currencyAmountIn, currencyOut, {
-//           maxHops: 3,
-//           maxNumResults: 1,
-//         });
+  useEffect(() => {
+    const getBestTradeExactIn = async () => {
+      if (currencyAmountIn && currencyOut && allowedPools.length > 0 && !isLoading) {
+        const trades = await ElixirTrade.bestTradeExactIn(allowedPools, currencyAmountIn, currencyOut, {
+          maxHops: 3,
+          maxNumResults: 1,
+        });
 
-//         const finalTrade = trades?.[0];
-//         setTradeData({ trade: finalTrade, isLoading: isLoading });
-//       }
-//     };
-//     getBestTradeExactIn();
-//   }, [allowedPools, isLoading, currencyAmountIn, currencyOut]);
-//   return tradeData;
-// }
+        const finalTrade = trades?.[0];
+        setTradeData({ trade: finalTrade, isLoading: isLoading });
+      }
+    };
+    getBestTradeExactIn();
+  }, [allowedPools, isLoading, currencyAmountIn, currencyOut]);
+  return tradeData;
+}
 
-// TODO: when add elixir package
 /**
  * Returns the best trade for the token in to the exact amount of token out
  */
-// export function useElixirTradeExactOut(
-//   currencyIn?: Currency,
-//   currencyAmountOut?: CurrencyAmount,
-// ): { trade: ElixirTrade | null; isLoading: boolean } {
-//   const [tradeData, setTradeData] = useState<{ trade: ElixirTrade | null; isLoading: boolean }>({
-//     trade: null,
-//     isLoading: true,
-//   });
+export function useElixirTradeExactOut(
+  currencyIn?: Currency,
+  currencyAmountOut?: CurrencyAmount,
+): { trade: ElixirTrade | null; isLoading: boolean } {
+  const [tradeData, setTradeData] = useState<{ trade: ElixirTrade | null; isLoading: boolean }>({
+    trade: null,
+    isLoading: true,
+  });
 
-//   const { pools: allowedPools, isLoading } = useAllElixirCommonPools(currencyIn, currencyAmountOut?.currency);
+  const { pools: allowedPools, isLoading } = useAllElixirCommonPools(currencyIn, currencyAmountOut?.currency);
 
-//   useEffect(() => {
-//     const getBestTradeExactOut = async () => {
-//       if (currencyIn && currencyAmountOut && allowedPools.length > 0 && !isLoading) {
-//         const trades = await ElixirTrade.bestTradeExactOut(allowedPools, currencyIn, currencyAmountOut, {
-//           maxHops: 3,
-//           maxNumResults: 1,
-//         });
+  useEffect(() => {
+    const getBestTradeExactOut = async () => {
+      if (currencyIn && currencyAmountOut && allowedPools.length > 0 && !isLoading) {
+        const trades = await ElixirTrade.bestTradeExactOut(allowedPools, currencyIn, currencyAmountOut, {
+          maxHops: 3,
+          maxNumResults: 1,
+        });
 
-//         const finalTrade = trades?.[0];
-//         setTradeData({ trade: finalTrade, isLoading: isLoading });
-//       }
-//     };
-//     getBestTradeExactOut();
-//   }, [allowedPools, isLoading, currencyIn, currencyAmountOut]);
-//   return tradeData;
-// }
+        const finalTrade = trades?.[0];
+        setTradeData({ trade: finalTrade, isLoading: isLoading });
+      }
+    };
+    getBestTradeExactOut();
+  }, [allowedPools, isLoading, currencyIn, currencyAmountOut]);
+  return tradeData;
+}
