+++ monorepo/packages/swap/src/hooks/useSwapCallback/hedera.ts
@@ -1,6 +1,6 @@
 import { BigNumber } from '@ethersproject/bignumber';
 import { CAVAX, JSBI, Percent, Router, Trade, TradeType } from '@pangolindex/sdk';
-import { INITIAL_ALLOWED_SLIPPAGE, useChainId, usePangolinWeb3, BIPS_BASE } from '@pangolindex/shared';
+import { BIPS_BASE, INITIAL_ALLOWED_SLIPPAGE, useChainId, usePangolinWeb3 } from '@pangolindex/shared';
 import { useTransactionAdder, useTransactionDeadline } from '@pangolindex/state-hooks';
 import { hederaFn } from '@pangolindex/wallet-connectors';
 import { useMemo } from 'react';
